requirements.txt
langchain
diffusers==0.16.1
fastapi==0.95.2
gradio==3.37.0
tiktoken==0.6.0
python-dotenv
openai
langchain_openai